Restore previously patched values if the rule was removed
If the rule is removed, and the user doesn't switch the theme,
then, when applying the settings, the value wouldn't be
touched at all, and would be stuck in the patched state.

To fix this issue, we save every value before patching
it for the first time, and then restore all of them before
patching again.

This logic assumes that the saved values remain
valid. This is not strictly true, as the values
could've been modified by some factor,
like the scaling factor or compact mode.
We'll have to fix this later.