Compact Mode support
If the compact mode is on, and the ruleset
specifies a ".compact" value,
use it instead of the regular one.

It would be easier if we could just
patch all values, including the ".compact" ones,
but unfortunately the IJ Platform
sends us a callback when the compact
values are already applied by replacing
the regular ones with them.