Improve theme selection UI
Remove the up/down actions.
Sort the themes alphabetically.

Fix the add action icon: a dropdown layer
was automatically added to it because
it's an action group, and it would be layered
on top of the existing one.

The list model implements using
whatever order the themes are given to it
instead of sorting inside the model
because later we can change the order.
Right now, the platform doesn't provide
the themes in the same order they're listed
in other places. This may change.